% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_network_metrics}
\alias{add_network_metrics}
\title{Add standard metrics to network collection}
\usage{
add_network_metrics(x, threshold = NULL)
}
\arguments{
\item{x}{Network collection.}

\item{threshold}{Optional probability threshold.}
}
\value{
A \code{network_collection} with summary columns appended to \code{index}; see \code{\link[=summarize_networks]{summarize_networks()}}.
}
\description{
Add standard metrics to network collection
}
