% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{list_shapefiles}
\alias{list_shapefiles}
\title{List ESRI shapefiles}
\usage{
list_shapefiles(path, recursive = FALSE)
}
\arguments{
\item{path}{Directory.}

\item{recursive}{Search recursively.}
}
\value{
A character vector of shapefile paths, empty if no matches are found.
}
\description{
List ESRI shapefiles
}
