% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaweave-package.R
\docType{package}
\name{metaweave-package}
\alias{metaweave}
\alias{metaweave-package}
\title{MetaWeave: spatial ecological network inference}
\description{
Reconstruct local assemblages from distributions, infer ecological networks
with pluggable models, simulate observations, summarize network properties,
and map the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fgabriel1891/metaweave}
  \item Report bugs at \url{https://github.com/fgabriel1891/metaweave/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Munoz \email{gmunozacevedo@proton.me} (Affiliation: Concordia University) [copyright holder]

Authors:
\itemize{
  \item Gabriel Munoz \email{gmunozacevedo@proton.me} (Affiliation: Concordia University) [copyright holder]
}

}
\keyword{internal}
