% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_network_collection}
\alias{new_network_collection}
\title{Construct a collection of spatial ecological networks}
\usage{
new_network_collection(networks, index, template = NULL, settings = list())
}
\arguments{
\item{networks}{List of ecological networks.}

\item{index}{Cell-level data frame.}

\item{template}{Spatial raster template.}

\item{settings}{Optional workflow settings.}
}
\value{
A \code{network_collection} list with \code{networks}, \code{index}, \code{template}, and \code{settings}. Each index row corresponds to one network.
}
\description{
Construct a collection of spatial ecological networks
}
