% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_spatial_result}
\alias{new_spatial_result}
\title{Construct a spatial result}
\usage{
new_spatial_result(data, metrics = names(data))
}
\arguments{
\item{data}{A spatial object.}

\item{metrics}{Names of mapped metrics.}
}
\value{
A \code{spatial_result} list with the spatial object in \code{data} and metric names in \code{metrics}.
}
\description{
Construct a spatial result
}
