% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{assemble_communities}
\alias{assemble_communities}
\title{Assemble local communities from aligned distribution rasters}
\usage{
assemble_communities(distributions, min_species = 1L)
}
\arguments{
\item{distributions}{Named list of SpatRaster or distribution_collection objects.}

\item{min_species}{Minimum species per group.}
}
\description{
Assemble local communities from aligned distribution rasters
}
