% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{block_model}
\alias{block_model}
\title{Construct a generic stochastic block model backend}
\usage{
block_model(
  row_lookup,
  column_lookup,
  theta,
  row_group,
  column_group,
  species_col = "species",
  guild_col = "guild"
)
}
\arguments{
\item{row_lookup}{Data frame mapping species in the matrix rows to blocks.}

\item{column_lookup}{Data frame mapping species in the matrix columns to blocks.}

\item{theta}{Guild-by-guild probability matrix.}

\item{row_group}{Name of the assemblage group placed in matrix rows.}

\item{column_group}{Name of the assemblage group placed in matrix columns.}

\item{species_col}{Species column in lookup tables.}

\item{guild_col}{Guild column in lookup tables.}
}
\description{
Construct a generic stochastic block model backend
}
