% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{downscale_networks}
\alias{downscale_networks}
\title{Compatibility wrapper for spatial network downscaling}
\usage{
downscale_networks(
  palm_stack,
  mammal_stack,
  palm_lookup,
  mammal_lookup,
  theta,
  species_col = "species",
  guild_col = "guild",
  min_palms = 1L,
  min_mammals = 1L
)
}
\description{
Compatibility wrapper for spatial network downscaling
}
