% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{infer_network}
\alias{infer_network}
\title{Infer a network for one assemblage}
\usage{
infer_network(assemblage, model)
}
\arguments{
\item{assemblage}{An assemblage.}

\item{model}{An inference model.}
}
\description{
Infer a network for one assemblage
}
