% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{infer_networks}
\alias{infer_networks}
\title{Infer networks for many assemblages}
\usage{
infer_networks(assemblages, model, template = NULL)
}
\arguments{
\item{assemblages}{List of assemblages.}

\item{model}{Inference model.}

\item{template}{Optional spatial template.}
}
\description{
Infer networks for many assemblages
}
