% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{interaction_table_to_matrix}
\alias{interaction_table_to_matrix}
\title{Convert an interaction table to a matrix}
\usage{
interaction_table_to_matrix(data, row_col, column_col, value_col, fill = 0)
}
\arguments{
\item{data}{Data frame.}

\item{row_col, column_col, value_col}{Column names.}

\item{fill}{Missing-pair value.}
}
\description{
Convert an interaction table to a matrix
}
