% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{make_bbox}
\alias{make_bbox}
\title{Create a bounding-box polygon}
\usage{
make_bbox(extent, crs = "EPSG:4326")
}
\arguments{
\item{extent}{Numeric vector \code{c(xmin, xmax, ymin, ymax)}.}

\item{crs}{Coordinate reference system.}
}
\description{
Create a bounding-box polygon
}
