% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{map_metrics}
\alias{map_metrics}
\title{Map cell metrics back to rasters}
\usage{
map_metrics(x, metrics = NULL)
}
\arguments{
\item{x}{Network collection with metrics in its index.}

\item{metrics}{Metric columns; inferred when omitted.}
}
\description{
Map cell metrics back to rasters
}
