% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{network_metrics_raster}
\alias{network_metrics_raster}
\title{Compatibility alias returning metric rasters}
\usage{
network_metrics_raster(x, metrics = NULL)
}
\description{
Compatibility alias returning metric rasters
}
