% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_assemblage}
\alias{new_assemblage}
\title{Construct a local assemblage}
\usage{
new_assemblage(species, cell_id = NA_integer_, coordinates = NULL)
}
\arguments{
\item{species}{Named list of character vectors, one per ecological group.}

\item{cell_id}{Optional raster cell identifier.}

\item{coordinates}{Optional named numeric vector containing x and y.}
}
\description{
Construct a local assemblage
}
