% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_distribution_collection}
\alias{new_distribution_collection}
\title{Construct a distribution collection}
\usage{
new_distribution_collection(data, guild = NULL, metadata = NULL)
}
\arguments{
\item{data}{A named \code{terra::SpatRaster} whose layers represent species.}

\item{guild}{Optional guild label for the collection.}

\item{metadata}{Optional species metadata data frame.}
}
\description{
Construct a distribution collection
}
