% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_ecological_network}
\alias{new_ecological_network}
\title{Construct an ecological network}
\usage{
new_ecological_network(
  matrix,
  type = c("probability", "binary", "weighted", "observed"),
  metadata = list(),
  row_group = "rows",
  column_group = "columns"
)
}
\arguments{
\item{matrix}{Numeric interaction matrix.}

\item{type}{One of probability, binary, weighted, or observed.}

\item{metadata}{Optional metadata.}
}
\description{
Construct an ecological network
}
