% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_inference_model}
\alias{new_inference_model}
\title{Construct a pluggable inference model}
\usage{
new_inference_model(
  predict,
  name = "custom",
  parameters = list(),
  subclass = NULL
)
}
\arguments{
\item{predict}{A function accepting \verb{(model, assemblage)} and returning an ecological network.}

\item{name}{Model name.}

\item{parameters}{Model-specific parameters.}

\item{subclass}{Optional additional class name.}
}
\description{
Construct a pluggable inference model
}
