% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{prepare_range_stacks}
\alias{prepare_range_stacks}
\title{Prepare palm and mammal range stacks}
\usage{
prepare_range_stacks(
  palm_source,
  mammal_source,
  mammal_species_col,
  mammal_species_keep = NULL,
  template = create_standard_grid(),
  region = NULL,
  output_dir = NULL,
  touches = TRUE
)
}
\description{
Prepare palm and mammal range stacks
}
