% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rasterize_range_rows}
\alias{rasterize_range_rows}
\title{Rasterize rows grouped by species}
\usage{
rasterize_range_rows(
  x,
  template,
  species_col,
  species_keep = NULL,
  touches = TRUE
)
}
\arguments{
\item{x}{An sf or SpatVector object.}

\item{template}{Raster template.}

\item{species_col}{Species-name column.}

\item{species_keep}{Optional species subset.}

\item{touches}{Count all touched cells.}
}
\description{
Rasterize rows grouped by species
}
