% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{simulate_interaction_counts}
\alias{simulate_interaction_counts}
\title{Simulate an interaction-count matrix}
\usage{
simulate_interaction_counts(network, size = 100L, seed = NULL)
}
\arguments{
\item{network}{Ecological network or probability matrix.}

\item{size}{Total number of interactions.}

\item{seed}{Optional random seed.}
}
\description{
Simulate an interaction-count matrix
}
