% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{simulate_network}
\alias{simulate_network}
\title{Simulate a network}
\usage{
simulate_network(network, size = 100L, seed = NULL)
}
\arguments{
\item{network}{Probability network.}

\item{size}{Total interaction count.}

\item{seed}{Optional seed.}
}
\description{
Simulate a network
}
