% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{summarize_network}
\alias{summarize_network}
\title{Summarize one ecological network}
\usage{
summarize_network(network, threshold = NULL)
}
\arguments{
\item{network}{Ecological network or numeric matrix.}

\item{threshold}{Optional threshold for realized links.}
}
\description{
Summarize one ecological network
}
