% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{summarize_networks}
\alias{summarize_networks}
\title{Summarize a collection of networks}
\usage{
summarize_networks(x, threshold = NULL)
}
\arguments{
\item{x}{Network collection.}

\item{threshold}{Optional probability threshold.}
}
\description{
Summarize a collection of networks
}
